+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * feature: `props.sortConfig.defaultSort` 添加 `silent` 属性，表示是否触发 `sort-change`事件。默认false。表示触发。
     - fix: 此配置用于修复defaultSort不触发sort-change 的问题。
 * fix: package.json types
+* fix: `highlightDimRow` 在 vue2.7 不生效的问题。
 
 ## 0.2.5
 * feature: `props.sortConfig.stringLocaleCompare` 控制是否使用 Array.prototype.localeCompare 排序字符串。默认true。